move detect pkf into sbom pkg
