+++ pkg/edit/utils.go
@@ -18,29 +18,14 @@ package edit
 
 import (
 	"errors"
-	"os"
 	"time"
-
-	"github.com/interlynk-io/sbomasm/pkg/detect"
 )
 
-var errNoConfiguration = errors.New("no configuration provided")
-var errNotSupported = errors.New("not supported")
-var errInvalidInput = errors.New("invalid input data")
-
-func detectSbom(path string) (string, string, error) {
-	f, err := os.Open(path)
-	if err != nil {
-		return "", "", err
-	}
-	defer f.Close()
-
-	spec, format, err := detect.Detect(f)
-	if err != nil {
-		return "", "", err
-	}
-	return string(spec), string(format), nil
-}
+var (
+	errNoConfiguration = errors.New("no configuration provided")
+	errNotSupported    = errors.New("not supported")
+	errInvalidInput    = errors.New("invalid input data")
+)
 
 func utcNowTime() string {
 	location, _ := time.LoadLocation("UTC")
